Restructure server utilities
