+++ tests/test_dispatch_loop.py
@@ -19,7 +19,7 @@ install_stubs()
 
 import main
 import orchestrator_agent
-import redis_manager
+from utils import redis_manager
 
 class FakeRedis:
     def __init__(self):
