+++ tests/test_import_hashes.py
@@ -18,7 +18,7 @@ install_stubs()
 
 
 import main
-import redis_manager
+from utils import redis_manager
 
 class FakeUploadFile:
     def __init__(self, name, data):
