+++ tests/test_process_hashes_jobs.py
@@ -20,7 +20,7 @@ install_stubs()
 
 
 import main
-import redis_manager
+from utils import redis_manager
 from uuid import UUID
