+++ tests/test_server_get_batch.py
@@ -16,7 +16,7 @@ install_stubs()
 
 
 import main
-import redis_manager
+from utils import redis_manager
 
 
 class FakeRedis:
