+++ tests/test_server_import_hashes.py
@@ -19,7 +19,7 @@ install_stubs()
 
 
 import main
-import redis_manager
+from utils import redis_manager
 from uuid import UUID
 
 class FakeUploadFile:
